Move exchange rate actions to a properly-named file
